chore(arena): move route dependencies to its own file
